+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Core: some modules couldn't be uninstalled from system settings
 - Core: Fixed too large cc field in e-mail composer
 - Core: Disabled phone number auto linking
+- Email: more user fliendly icon and tooltip for CC / BCC submenu
 
 13-10-2020 6.4.181
 - Core: In system settings, extra filter 'Disabled users', make sortable by change date